requirements.txt
accelerate==0.23.0
bitsandbytes==0.41.1
bert-score
datasets==2.13.0
evaluate==0.4.0
fasttext-wheel==0.9.2
gpustat
gradio
gradio-client
huggingface-hub
langchain==0.0.310
matplotlib
openai
peft==0.5.0
ray
rouge_score
sacrebleu
sacremoses
safetensors>=0.3.1
seaborn
tokenizers
transformers==4.34.0
trl==0.7.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tqdm
wandb==0.15.4
